requirements.txt
snowflake-connector-python>=3.12.0
numpy>=1.26.0
pandas>=2.1.0
nba_api>=1.4.1
python-dotenv>=1.0.0 